change structure controllers
