+++ controllers/avatarController/index.js
@@ -1,0 +1,3 @@
+const updateAvatar = require("./avatarController");
+
+module.exports = { updateAvatar };
