+++ controllers/usersControllers/currentController.js
@@ -1,0 +1,6 @@
+const currentController = async (req, res, next) => {
+  const { email } = req.user;
+  res.json({ status: "success", email });
+};
+
+module.exports = currentController;
