+++ controllers/usersControllers/index.js
@@ -1,0 +1,13 @@
+const signupController = require("./signupController");
+const loginController = require("./loginController");
+const logoutController = require("./logoutController");
+const currentController = require("./currentController");
+const changeSubscriptionController = require("./changeSubscriptionController");
+
+module.exports = {
+  signupController,
+  loginController,
+  logoutController,
+  currentController,
+  changeSubscriptionController,
+};
