+++ db/index.js
@@ -1,0 +1,5 @@
+const User = require("./usersModel");
+const connectMongo = require("./connectMongo");
+const Contact = require("./contactModel");
+
+module.exports = { User, connectMongo, Contact };
