+++ db/usersModel.js
@@ -33,6 +33,4 @@ userSchema.pre("save", async function () {
 
 const User = mongoose.model("user", userSchema);
 
-module.exports = {
-  User,
-};
+module.exports = { User };
